Redo how we display defaults.
As the to_json and the default changes I was working on (But didn't commit) had
a lot in common, I decided to use the to_json methods to fill in the default
value. This saves us from replicating this code twice per class.
I also chose to have the default property not appear at all if there is no
default present as anything else can be a valid default value.